[goal:module-split-calculators] feat(assets): add calculator modules
