+++ assets/js/hero-optimizer.js
@@ -1,0 +1,20 @@
+(function() {
+    function initHeroOptimizer() {
+        // Placeholder for hero optimization calculator
+        // Original implementation relied on yet-to-be-defined helper methods.
+    }
+
+    if (typeof document !== 'undefined') {
+        document.addEventListener('DOMContentLoaded', () => {
+            if (document.getElementById('heroOptimizer')) {
+                initHeroOptimizer();
+            }
+        });
+    }
+
+    if (typeof module !== 'undefined') {
+        module.exports = { initHeroOptimizer };
+    } else {
+        window.initHeroOptimizer = initHeroOptimizer;
+    }
+})();
