+++ assets/js/resource-efficiency.js
@@ -1,0 +1,20 @@
+(function() {
+    function initResourceEfficiency() {
+        // Placeholder for resource efficiency calculator
+        // Original implementation relied on yet-to-be-defined helper methods.
+    }
+
+    if (typeof document !== 'undefined') {
+        document.addEventListener('DOMContentLoaded', () => {
+            if (document.getElementById('resourceEfficiency')) {
+                initResourceEfficiency();
+            }
+        });
+    }
+
+    if (typeof module !== 'undefined') {
+        module.exports = { initResourceEfficiency };
+    } else {
+        window.initResourceEfficiency = initResourceEfficiency;
+    }
+})();
