+++ pages/T10-calculator.html
@@ -188,7 +188,9 @@
     </script>
   <script src="../assets/js/storage-utils.js"></script>
   <script src="../assets/js/script.js"></script>
-  <script src="../assets/js/calculators.js"></script>
+  <script type="module">
+    import('../assets/js/t10-calculator.js');
+  </script>
   <script src="../assets/js/analytics.js"></script>
 </body>
